fix(trusty-installer): bound the verify-tail fallback re-probe + inject ServiceEnv
Code-critic fix round on PR #3849 (APPROVE, 2 MEDIUM):

- MEDIUM 1: the #3841 verify-tail bootstrap-fallback re-probed health
  exactly once, instantly, after a successful fallback -- racing a
  just-repaired but slow-starting daemon (trusty-search's first-boot
  embedding-model download) into a false down/NOT VERIFIED, the same
  race class #3833 already fixed for the kickstart path. The
  post-fallback re-probe now goes through the same bounded
  poll_until_not_down/bounded_attempts machinery, with the remaining
  budget recomputed from elapsed wall-clock so it correctly reflects
  whatever the member's own kickstart phase already spent.

- MEDIUM 2: attempt_verify_fallback hardcoded RealServiceEnv, so the
  entire NotLoaded -> fallback -> re-poll/re-classify wiring had zero
  unit coverage. Extracted as apply_not_loaded_fallback, taking
  &dyn ServiceEnv (mirroring bootstrap_one's own seam), parameterized
  to RealServiceEnv only at the real call site
  (apply_not_loaded_fallback_real). Four new unit tests cover both
  outcomes plus the no-plist and exhausted-budget edge cases.

verify_tail.rs crossed the 500-SLOC production cap once these fixes
plus their tests landed; tests move to the sibling verify_tail_tests.rs
(mirrors the install.rs / install_tests.rs split).

Refs #3848, #3849
